fix: Clippy warnings and SceneNode parent field in tests/examples
This commit addresses several issues discovered during Phase 3 final integration:

- Add missing `parent: None` field to SceneNode struct literals in:
  - arthropod-ecs tests (ecs_tests.rs, ecs_resource_tests.rs)
  - arthropod tests (app_builder_tests.rs)
  - examples (colored_rectangles.rs, debug_render.rs, single_rect_test.rs)

- Fix clippy warnings:
  - text-engine: remove unnecessary `.into()` and `as u16` casts
  - render-engine: remove unnecessary `as u32` casts, use if-let instead of match
  - arthropod: add collapsible_if allow attributes (nested if-let is more readable),
    fix needless_borrow, add WidgetBuilder type alias for complexity

All tests pass and clippy runs clean.

+++ crates/render-engine/src/backend/text/glyph_atlas.rs
@@ -1,7 +1,7 @@
 //! Glyph atlas for caching rasterized glyphs in a texture
 
-use hashbrown::HashMap;
 use cosmic_text::{CacheKey, FontSystem, SwashCache};
+use hashbrown::HashMap;
 
 /// Texture coordinates in atlas (0.0-1.0 normalized)
 #[derive(Debug, Clone, Copy, PartialEq)]
@@ -87,8 +87,8 @@ impl GlyphAtlas {
         let (glyph_width, glyph_height, glyph_data) =
             match self.swash_cache.get_image(font_system, cache_key) {
                 Some(img) => {
-                    let w = img.placement.width as u32;
-                    let h = img.placement.height as u32;
+                    let w = img.placement.width;
+                    let h = img.placement.height;
                     let data = img.data.clone(); // Clone to release borrow on self
                     (w, h, data)
                 }
